+++ src/main/java/com/chukchuk/haksa/application/portal/SyncAcademicRecordService.java
@@ -11,8 +11,8 @@ import com.chukchuk.haksa.domain.course.model.CourseOffering;
 import com.chukchuk.haksa.domain.course.service.CourseOfferingService;
 import com.chukchuk.haksa.domain.course.service.CourseService;
 import com.chukchuk.haksa.domain.professor.service.ProfessorService;
-import com.chukchuk.haksa.domain.student.model.Grade;
-import com.chukchuk.haksa.domain.student.model.GradeType;
+import com.chukchuk.haksa.domain.student.model.grade.Grade;
+import com.chukchuk.haksa.domain.student.model.grade.GradeType;
 import com.chukchuk.haksa.domain.student.model.Student;
 import com.chukchuk.haksa.domain.student.service.StudentService;
 import com.chukchuk.haksa.infrastructure.portal.mapper.AcademicRecordMapperFromPortal;
